requirements.txt

google-generativeai
python_dotenv
fastapi
uvicorn
requests
validators
scikit-learn
pandas